Move CubeMap loading logic to XwaD3dRendererHook.cpp
... that way we can load the map while a mission is loading.
